[Refactor] Component Structured
